requirements.txt
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
biopython>=1.79
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.65.0
pillow>=9.0.0
scipy>=1.7.0
tensorboard>=2.12.0
einops>=0.8.1

